+++ CHANGES.md
@@ -1,3 +1,4 @@
 # 0.1.0
 * First version
+* Support for interacting with events
 * Support for interacting with stashes